move smaller chat components to app ns
