+++ src/app/chat/Badge.vue
