+++ src/app/chat/BadgeTooltip.vue
