+++ src/app/chat/EmoteTooltip.vue
