+++ src/app/chat/MessageTokenLink.vue
