+++ src/app/chat/RichEmbed.vue
