+++ src/app/chat/types/0.NormalMessage.vue
