+++ src/app/chat/types/11.Connected.vue
