+++ src/app/chat/types/43.PointsReward.vue
