+++ src/app/chat/types/49.AnnouncementMessage.vue
