+++ src/app/chat/types/BasicSystemMessage.vue
