+++ src/site/twitch.tv/modules/chat/ChatData.vue
@@ -8,7 +8,7 @@ import { useChatEmotes } from "@/composable/chat/useChatEmotes";
 import { useChatMessages } from "@/composable/chat/useChatMessages";
 import { useLiveQuery } from "@/composable/useLiveQuery";
 import { WorkletEvent, useWorker } from "@/composable/useWorker";
-import EmoteSetUpdateMessage from "./components/types/EmoteSetUpdateMessage.vue";
+import EmoteSetUpdateMessage from "@/app/chat/types/EmoteSetUpdateMessage.vue";
 import { v4 as uuidv4 } from "uuid";
 
 const { target } = useWorker();
